+++ scripts/start_functions.R
@@ -377,13 +377,15 @@ start_run <- function(cfg, scenario = NULL, codeCheck = TRUE, lock_model = TRUE,
     ad_aolc_pol    <- magclass::read.magpie("modules/35_natveg/input/npi_ndc_ad_aolc_pol.cs3")
     ad_pol         <- ad_aolc_pol[,,"forest"]
     aolc_pol       <- ad_aolc_pol[,,"other"]
-    affexp_missing <- (cfg$gms$c32_aff_policy == "affexp") && 
+    affexp_missing <- (cfg$gms$c32_aff_policy == "affexp") &&
                         !("affexp" %in% magclass::getNames(aff_pol))
+    ndcdelay_missing <- (cfg$gms$c32_aff_policy == "ndcdelay") &&
+                        !("ndcdelay" %in% magclass::getNames(aff_pol))
 
-    if((all(aff_pol == 0)  && (cfg$gms$c32_aff_policy != "none"))  || 
+    if((all(aff_pol == 0)  && (cfg$gms$c32_aff_policy != "none"))  ||
        (all(ad_pol == 0)   && (cfg$gms$c35_ad_policy != "none"))   ||
        (all(aolc_pol == 0) && (cfg$gms$c35_aolc_policy != "none")) ||
-       affexp_missing)
+       affexp_missing || ndcdelay_missing)
     {
       cfg$recalc_npi_ndc <- TRUE
     } else cfg$recalc_npi_ndc <- FALSE